Remove a dereference.  It was unintended and a no-op.
Use the correct value of errno.  Although the errno value passed into
printf() follows the *env() call, it is not guaranteed to be the errno
from that call.  When I wrote the regression tester, the environment I
used did pass the errno from the call.  Consolidate the print for the
return code and errno into a function in the process of fixing this.

Approved by:	wes (mentor)


git-svn-id: svn://svn.freebsd.org/base/head@176676 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f