Add a -n option to pkg_create(1), to inhibit duplicate work.
When run without this option, multiple runs of `pkg_create -Rb' will
recreate common packages multiple times.  This can take a lot of time
for large packages.  With the -n option `pkg_create -b' checks with
stat(2) and skips packages that already exist.

Note that this may *not* be safe of the existing output file is not
really a package, or if it has been corrupted, modified or otherwise
tinkered with between subsequent pkg_create runs.  For this and POLA
reasons, the default behavior is to *rebuild* the packages, and the -n
option can be used when we know it is `safe' to run in no-regenerate
mode.

Inspired by:	A post to freebsd-questions
	 	by Matthias Apitz &lt; matthias.apitz at oclc.org &gt;
Reviewed by:	marcus, flz
Approved by:	marcus
MFC after:	2 weeks


git-svn-id: svn://svn.freebsd.org/base/head@179352 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f